+++ src/components/Messaging/MessageBody.js
@@ -1,17 +1,36 @@
 import React, { useState, useContext, useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core/styles';
-import { Avatar, Grid, Typography, Paper } from '@material-ui/core';
+import { Announcement } from '@material-ui/icons';
+import { Avatar, Grid, Typography, Paper, Button } from '@material-ui/core';
 import { TextInput } from './TextInput.js';
+import dateFormat from 'dateformat';
 
 import { MessagingContext } from 'context/MessagingContext.js';
 
 const useStyles = makeStyles((theme) => ({
   messageRow: {
     display: 'flex',
+    padding: '3px',
   },
   messageRowRight: {
     display: 'flex',
     justifyContent: 'flex-end',
+  },
+  messageHeader: {
+    display: 'flex',
+    borderBottom: '.3px solid #fff',
+    marginBottom: '5px',
+  },
+  announceMessage: {
+    position: 'relative',
+    marginLeft: '20px',
+    marginBottom: '10px',
+    padding: '10px',
+    background: theme.palette.primary.main,
+    color: 'white',
+    textAlign: 'left',
+    borderRadius: '10px',
   },
   recievedMessage: {
     position: 'relative',
@@ -27,9 +46,15 @@ const useStyles = makeStyles((theme) => ({
     marginRight: '20px',
     marginBottom: '10px',
     padding: '10px',
-    backgroundColor: theme.palette.primary.main,
+    backgroundColor: theme.palette.primary.lightMed,
     textAlign: 'left',
     borderRadius: '10px',
+  },
+  messageTitle: {
+    padding: 0,
+    margin: 0,
+    wordWrap: 'break-word',
+    fontWeight: 'bold',
   },
   messageContent: {
     padding: 0,
@@ -39,7 +64,7 @@ const useStyles = makeStyles((theme) => ({
   messageTimeStampLeft: {
     display: 'flex',
     marginRight: 'auto',
-    alignItems: 'center',
+    alignItems: 'top',
     color: 'grey',
     [theme.breakpoints.down('xs')]: {
       display: 'none',
@@ -68,7 +93,7 @@ const useStyles = makeStyles((theme) => ({
     borderBottom: '2px solid black',
   },
   messagesBody: {
-    height: '80%',
+    flexGrow: 1,
     width: '95%',
     padding: '7.5px',
     overflowY: 'auto',
@@ -88,29 +113,68 @@ const useStyles = makeStyles((theme) => ({
       width: '2em',
       height: '2em',
     },
+    [theme.breakpoints.down('xs')]: {
+      display: 'none',
+    },
   },
   textInput: {
     borderTop: '2px solid black',
   },
   surveyContent: {
-    color: '#fff',
+    display: 'flex',
+    flexDirection: 'column',
+    marginLeft: 'auto',
+    backgroundColor: theme.palette.primary.lightMed,
+    // backgroundColor: theme.palette.primary.main,
+    // color: '#fff',
+    width: '65%',
+    borderRadius: '10px',
+    padding: '10px',
+  },
+  dataContainer: {
+    marginLeft: 'auto',
+    alignSelf: 'center',
+  },
+  button: {
+    backgroundColor: theme.palette.primary.main,
+    borderRadius: '50px',
+    color: 'white',
+    margin: '5px',
+  },
+  surveyResponse: {
+    display: 'flex',
+    flexDirection: 'column',
+    backgroundColor: 'lightGrey',
+    borderRadius: '10px',
+    padding: '1em',
+    margin: '5px',
+    width: '35%',
   },
 }));
 
 export const AnnounceMessage = ({ message }) => {
   const {
     messageRow,
-    recievedMessage,
+    messageHeader,
+    messageTitle,
+    announceMessage,
     messageContent,
     messageTimeStampRight,
   } = useStyles();
 
   return (
     <div className={messageRow}>
-      <div className={recievedMessage}>
-        <div>
-          <Typography className={messageContent}>{message.body}</Typography>
+      <div className={announceMessage}>
+        <div className={messageHeader}>
+          <Announcement color="inherit" style={{ padding: '2px' }} />
+          <Typography className={messageTitle}>{message.title}</Typography>
         </div>
+        <Typography className={messageContent}>{message.body}</Typography>
+        {message.video_link && (
+          <Typography className={messageContent} variant="body1">
+            {message.video_link}
+          </Typography>
+        )}
       </div>
       <Grid item className={messageTimeStampRight}>
         <Typography>{message.composed_at.slice(0, 10)}</Typography>
@@ -119,36 +183,67 @@ export const AnnounceMessage = ({ message }) => {
   );
 };
 
-export const SurveyMessage = ({ message }) => {
-  const { messageRowRight, sentMessage, surveyContent } = useStyles();
+export const SurveyResponseMessage = ({ message }) => {
+  const { messageRow, messageTimeStampRight, surveyResponse } = useStyles();
 
-  const { questions } = message.survey;
   return (
-    <div className={messageRowRight}>
-      <Grid item className={sentMessage}>
-        <Typography variant={'h5'} className={surveyContent}>
-          {message.body ? message.body : ''}
-        </Typography>
-        {questions.map((question, i) => (
-          <div key={question + `:${i + 1}`} className={surveyContent}>
-            <Typography variant={'h6'}>
-              Question {i + 1}:{' '}
-              <Typography variant={'body1'}>{question.prompt}</Typography>
-            </Typography>
-            <Typography variant={'h6'}>
-              Choices:
-              <ol type="A">
-                {question.choices.map((choice, i) => (
-                  <li key={choice ? `choice ${i + 1}:${choice}` : i}>
-                    {choice}
-                  </li>
-                ))}
-              </ol>
-            </Typography>
-          </div>
-        ))}
+    <div className={messageRow}>
+      <Grid className={surveyResponse}>
+        {message.survey?.answers &&
+          message.survey.answers.map((answer, i) => (
+            <div key={`answer - ${i}`}>
+              <Typography variant={'h6'}>Question {i + 1}:</Typography>
+              <Typography variant={'body1'}>{answer}</Typography>
+            </div>
+          ))}
+      </Grid>
+      <Grid item className={messageTimeStampRight}>
+        <Typography>{message.composed_at.slice(0, 10)}</Typography>
       </Grid>
     </div>
+  );
+};
+
+export const SurveyMessage = ({ message }) => {
+  const { messageRow, surveyContent, messageTimeStampLeft } = useStyles();
+
+  return (
+    <>
+      {message.survey.response ? (
+        <SurveyResponseMessage message={message} />
+      ) : (
+        <div className={messageRow}>
+          <Grid item className={messageTimeStampLeft}>
+            <Typography>
+              {dateFormat(message.composed_at, 'yyyy-mm-dd hh:mm')}
+            </Typography>
+          </Grid>
+          <Grid item className={surveyContent}>
+            <Typography variant={'h4'}>
+              {message.body ? message.body : ''}
+            </Typography>
+            {message.survey.questions.map((question, i) => (
+              <div key={question + `:${i + 1}`}>
+                <Typography variant={'h6'}>
+                  Question {i + 1}:{' '}
+                  <Typography variant={'body1'}>{question.prompt}</Typography>
+                </Typography>
+                <Typography variant={'h6'}>
+                  Choices:
+                  <ol type="A">
+                    {question.choices.map((choice, i) => (
+                      <li key={choice ? `choice ${i + 1}:${choice}` : i}>
+                        {choice}
+                      </li>
+                    ))}
+                  </ol>
+                </Typography>
+              </div>
+            ))}
+          </Grid>
+        </div>
+      )}
+    </>
   );
 };
 
@@ -201,8 +296,8 @@ export const SentMessage = ({ message }) => {
   );
 };
 
-const SenderInformation = ({ messageRecipient, subject, id }) => {
-  const { senderInfo, senderItem, avatar } = useStyles();
+const SenderInformation = ({ message, messageRecipient, subject, id }) => {
+  const { senderInfo, senderItem, avatar, button, dataContainer } = useStyles();
 
   return (
     <Grid container className={senderInfo}>
@@ -213,15 +308,28 @@ const SenderInformation = ({ messageRecipient, subject, id }) => {
         <Typography variant="h5">
           {subject === 'Survey' ? subject : messageRecipient}
         </Typography>
-        <Typography align="left" color="primary" variant="h6">
+
+        {(subject === 'Survey' || subject === 'Announce') && (
+          <Typography>
+            DATE: {dateFormat(message?.composed_at, 'yyyy/mm/dd')}
+          </Typography>
+        )}
+
+        <Typography align="left" color="primary">
           {subject === 'Survey' ? messageRecipient : `ID: ${id}`}
         </Typography>
       </Grid>
+      {subject === 'Survey' && (
+        <Grid item className={dataContainer}>
+          <Button className={button}>Survey Data</Button>
+        </Grid>
+      )}
     </Grid>
   );
 };
 
 const MessageBody = ({ messages, messageRecipient }) => {
+  const history = useHistory();
   const { paper, messagesBody, textInput } = useStyles();
   const { user, authors, postMessageSend } = useContext(MessagingContext);
   const [messageContent, setMessageContent] = useState('');
@@ -237,13 +345,14 @@ const MessageBody = ({ messages, messageRecipient }) => {
   useEffect(() => {
     if (authors && messageRecipient) {
       let res = authors.find((author) => author.handle === messageRecipient);
+
       if (res) {
         setRecipientId(res.id);
       }
     }
-  }, [messageRecipient]);
+  }, [authors, messageRecipient]);
 
-  const handleSubmit = (e) => {
+  const handleSubmit = async (e) => {
     e.preventDefault();
 
     let lastMessage = messages[messages.length - 1];
@@ -258,7 +367,8 @@ const MessageBody = ({ messages, messageRecipient }) => {
 
     if (messageContent !== '') {
       if (user.userName && messageRecipient) {
-        postMessageSend(messagePayload);
+        await postMessageSend(messagePayload);
+        history.go(0);
       }
     }
     setMessageContent('');
@@ -268,6 +378,7 @@ const MessageBody = ({ messages, messageRecipient }) => {
     <Paper className={paper}>
       {messageRecipient && messages ? (
         <SenderInformation
+          message={messages[0]}
           messageRecipient={messageRecipient}
           subject={subject}
           id={recipientId}
@@ -279,7 +390,7 @@ const MessageBody = ({ messages, messageRecipient }) => {
         {messages ? (
           messages.map((message, i) => {
             if (message.subject === 'Message') {
-              return message.from === user.userName ? (
+              return message.from.author === user.userName ? (
                 <SentMessage
                   key={message.id ? `messageId=${message.id}i=${i}` : `i`}
                   message={message}
@@ -312,13 +423,15 @@ const MessageBody = ({ messages, messageRecipient }) => {
           <div>Loading ...</div>
         )}
       </div>
-      <TextInput
-        messageRecipient={messageRecipient}
-        handleSubmit={handleSubmit}
-        messageContent={messageContent}
-        setMessageContent={setMessageContent}
-        className={textInput}
-      />
+      {subject !== 'Survey' && (
+        <TextInput
+          messageRecipient={messageRecipient}
+          handleSubmit={handleSubmit}
+          messageContent={messageContent}
+          setMessageContent={setMessageContent}
+          className={textInput}
+        />
+      )}
     </Paper>
   );
 };