+++ src/components/Messaging/Messaging.js
@@ -1,6 +1,5 @@
 import React, { useState, useEffect, useContext } from 'react';
 
-import Navbar from 'components/Navbar';
 import Inbox from './Inbox';
 import MessageBody from './MessageBody';
 import Survey from './Survey';
@@ -12,13 +11,12 @@ import { makeStyles } from '@material-ui/styles';
 import { MessagingContext } from '../../context/MessagingContext';
 
 const useStyles = makeStyles((theme) => ({
-  rootContainer: {
-    paddingBottom: '1em',
-    marginBottom: '2em',
-  },
-  title: {
-    marginLeft: theme.spacing(7),
-    fontSize: '24px',
+  headerGrid: {
+    width: '90%',
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    margin: 'auto',
   },
   button: {
     backgroundColor: theme.palette.primary.main,
@@ -26,22 +24,19 @@ const useStyles = makeStyles((theme) => ({
     color: 'white',
     margin: '5px',
   },
-  buttonContainer: {
-    margin: '2em',
-    marginLeft: 'auto',
-  },
   messagesContainer: {
     margin: '2em',
     padding: '2em',
   },
   container: {
     width: '90vw',
-    height: '85vh',
     marginLeft: 'auto',
     marginRight: 'auto',
-    marginBottom: '6em',
+    marginBottom: '5vw',
     border: '1px solid black',
     borderRadius: '5px',
+    display: 'flex',
+    flexGrow: 1,
   },
   inbox: {
     width: '30%',
@@ -57,29 +52,14 @@ const useStyles = makeStyles((theme) => ({
 
 const Messaging = () => {
   // styles
-  const {
-    rootContainer,
-    title,
-    button,
-    buttonContainer,
-    container,
-    inbox,
-    body,
-  } = useStyles();
-
+  const { headerGrid, button, container, inbox, body } = useStyles();
   const { user, messages, loadMessages, loadRegions, loadAuthors } = useContext(
     MessagingContext
   );
 
-  useEffect(() => {
-    loadMessages();
-    loadRegions();
-    loadAuthors();
-  }, []);
-
   const [toggleAnnounceMessage, setToggleAnnounceMessage] = useState(false);
   const [toggleSurvey, setToggleSurvey] = useState(false);
-  const [messageRecipient, setMessageRecipient] = useState('');
+  const [messageRecipient, setMessageRecipient] = useState(null);
   const [selectedIndex, setSelectedIndex] = useState(0);
   const [openModal, setOpenModal] = useState(false);
   const handleOpen = () => setOpenModal(true);
@@ -88,11 +68,18 @@ const Messaging = () => {
   const findMessageRecipient = (messagesArray) => {
     return messagesArray[0].messages[0].to[0].recipient !== user.userName
       ? setMessageRecipient(messagesArray[0].messages[0].to[0].recipient)
-      : setMessageRecipient(messagesArray[0].messages[0].from);
+      : setMessageRecipient(messagesArray[0].messages[0].from.author);
   };
 
   useEffect(() => {
-    if (messages.length && messageRecipient === '') {
+    console.log('Messaging.js: useEffect: loadMessages');
+    loadMessages();
+    loadRegions();
+    loadAuthors();
+  }, []);
+
+  useEffect(() => {
+    if (messages.length && messageRecipient === null) {
       findMessageRecipient(messages);
     }
   }, [messages]);
@@ -104,12 +91,11 @@ const Messaging = () => {
 
   return (
     <>
-      <Navbar />
-      <Grid container direction="row" className={rootContainer} id="Messaging">
-        <Grid item className={title}>
+      <Grid container id="Messaging" className={headerGrid}>
+        <Grid item>
           <h1>Inbox</h1>
         </Grid>
-        <Grid item className={buttonContainer}>
+        <Grid item>
           <Button className={button} onClick={handleOpen}>
             New Message
           </Button>