+++ src/components/common/CustomTableFilter/CustomTableFilter.js
@@ -14,6 +14,7 @@ import Divider from '@material-ui/core/Divider';
 import CloseIcon from '@material-ui/icons/Close';
 import FormControl from '@material-ui/core/FormControl';
 import useStyles from './CustomTableFilter.styles';
+import { AppContext } from '../../../context/AppContext';
 
 const PAYMENT_STATUS = ['calculated', 'cancelled', 'paid', 'all'];
 
@@ -41,6 +42,7 @@ function CustomTableFilter(props) {
   } = props;
 
   const classes = useStyles();
+  const { updateSelectedFilter } = React.useContext(AppContext);
 
   const handleOnFormControlChange = (e) => {
     e.preventDefault();
@@ -60,19 +62,20 @@ function CustomTableFilter(props) {
     ); // filter out keys we don't want to submit
     setFilter(filtersToSubmit);
     setIsFilterOpen(false);
+    updateSelectedFilter(filtersToSubmit);
   };
 
   const handleOnFilterFormReset = (e) => {
     e.preventDefault();
     if (Object.keys(localFilter).length !== 0) {
-      const withoutMainFilter = Object.assign({}, filter);
-      delete withoutMainFilter.earnings_status;
-      delete withoutMainFilter.funder_id;
-      delete withoutMainFilter.grower;
-      delete withoutMainFilter.phone;
-      delete withoutMainFilter.start_date;
-      delete withoutMainFilter.end_date;
-      setFilter(withoutMainFilter);
+      const withoutLocalFilter = Object.assign({}, filter);
+      delete withoutLocalFilter.earnings_status;
+      delete withoutLocalFilter.funder_id;
+      delete withoutLocalFilter.grower;
+      delete withoutLocalFilter.phone;
+      delete withoutLocalFilter.start_date;
+      delete withoutLocalFilter.end_date;
+      setFilter(withoutLocalFilter);
       setLocalFilter({});
     }
     setIsFilterOpen(false);